tests: Split out test_invite from test_signup.
There is no good reason for this single test file to be more than 6k
lines.
